+++ include/Node.h
@@ -4,9 +4,9 @@
 
 struct Point {
 	Point() {};
-	Point(int x, int y) { this->x = x, this->y = y; };
-	int x;
-	int y;
+	Point(float x, float y) { this->x = x, this->y = y; };
+	float x;
+	float y;
 
 	friend bool operator==(const Point& p1, const Point& p2) {
 		if (p1.x == p2.x && p1.y == p2.y) return true;
